# Conflicts:
#	Podfile
#	Signal/Signal-Info.plist

+++ SignalServiceKit/Upload/AttachmentUploadManager.swift
@@ -287,7 +287,7 @@ public actor AttachmentUploadManagerImpl: AttachmentUploadManager {
         )
 
         let wrappedProgress: OWSProgressSink = OWSProgress.createSink { [weak self] progressValue in
-            self?.updateProgress(id: attachmentId, progress: Double(progressValue.percentComplete))
+            self?.updateProgress(id: attachmentId, progress: progressValue.percentComplete)
             if let progressSource, progressSource.completedUnitCount < progressValue.completedUnitCount {
                 progressSource.incrementCompletedUnitCount(
                     by: progressValue.completedUnitCount - progressSource.completedUnitCount,
@@ -1164,7 +1164,7 @@ public actor AttachmentUploadManagerImpl: AttachmentUploadManager {
         return try .validateAndBuild(fileUrl: tmpReencryptedFile, metadata: reencryptedMetadata)
     }
 
-    private nonisolated func updateProgress(id: Attachment.IDType, progress: Double) {
+    private nonisolated func updateProgress(id: Attachment.IDType, progress: Float) {
         NotificationCenter.default.postOnMainThread(
             name: Upload.Constants.attachmentUploadProgressNotification,
             object: nil,
